requirements.txt
Flask~=2.3.2
ipywidgets~=8.0.7
ipython~=8.14.0
gunicorn
torch~=2.0.1
diffusers~=0.17.1
requests~=2.31.0
Pillow~=9.5.0